Updated Gregorian Definitions
Moved calendar components enum out of Gregorian namespace
